Split QueryStoreGridControl.axaml.cs into partial classes (#330)
Move-only refactor; no behavior changes. QueryStoreGridControl.axaml.cs
(1,959 lines) split into 7 partials:

  Fetch       (452) - Fetch_Click + FetchPlans/WaitStats + grouping
  Filters     (256) - search type/value + column filter popup + ApplyFilters
  GroupBy     (161) - GroupBy + ReorderColumns + row expand/collapse
                      + ResultsGrid_DoubleTapped + ToggleRowExpansion
  Sort        (245) - sorting + UpdateBarRatios + GetSortKey + status
  Selection   (134) - select/load/ViewHistory/context menu + copy
  WaitStats   (182) - wait fetch + category click + mode toggle
  TimeRange   ( 33) - OnTimeRangeChanged

Main file now 465 lines — fields, ctor, public API
(SetInitialTimeRange), DB picker, and static accessor dictionaries.

The standalone QueryStoreRow class at the bottom of the file is
untouched.

Build clean: 0 errors, 0 warnings on PlanViewer.App.
